deletar resposta

+++ src/domain/forum/application/use-cases/delete-question/delete-question.ts
@@ -2,6 +2,7 @@ import { QuestionsRepository } from '../../repositories/questions-repository'
 
 interface DeleteQuestionUseCaseRequest {
  questionId: string
+ authorId : string
 }
 
 interface DeleteQuestionUseCaseResponse {
@@ -12,12 +13,17 @@ export class DeleteQuestionUseCase {
   constructor(private questionRepository: QuestionsRepository) {}
 
   async execute({
-    questionId
+    questionId,
+    authorId
   }: DeleteQuestionUseCaseRequest) : Promise<DeleteQuestionUseCaseResponse> {
     const question = await this.questionRepository.findById(questionId)
 
     if(!question){
         throw new Error('Question not found')
+    }
+
+    if(authorId !== question.authorId.toString()){
+        throw new Error('Not Alowed')
     }
 
    await this.questionRepository.delete(question)
